avoid pre-instantiating env setters
